+++ src/main.c
@@ -30,54 +30,23 @@ t_minishell	*init_minishell(char **envp)
 }
 
 /**
- * Main - Get size of readline's leak
- * Run `valgrind ./minishell`
+ * Reads input from the user and handles signal-related behavior.
  *
-int	main(void)
-{
-	char	*input;
-
-	while (1)
-	{
-		input = readline("readline_test$ ");
-		if (ft_strncmp("exit", input, 5) == 0)
-		{
-			ft_printf("exit\n");
-			ft_free(1, &input);
-			exit(EXIT_SUCCESS);
-		}
-		ft_printf("input: \"%s\"\n", input);
-		ft_free(1, &input);
-	}
-	return (EXIT_SUCCESS);
-}
-*/
-
-/**
- * Main - Debug version (skip readline to check leaks)
+ * - Prompts the user for input using `readline`.
+ * - Adds the input to the command history.
+ * - Handles SIGINT signals and updates the exit code accordingly.
+ * - Exits the minishell gracefully if EOF (Ctrl+D) is received.
  *
-int	main(int ac, char **av, char **envp)
-{
-	t_minishell	*minishell;
-	char		*input;
-
-	(void)av;
-	minishell = init_minishell(envp);
-	input = "<test/infile tail -n +4 | grep a | sort | uniq -c | sort -nr \
-		| head -n 3";
-	init_cmd_lst(input, minishell);
-	execute_cmd_lst(minishell);
-	free_minishell(&minishell);
-	return (EXIT_SUCCESS);
-}
-*/
-
+ * @param ms Pointer to the minishell data structure.
+ * @return void
+ */
 static void	set_input(t_minishell *ms)
 {
 	int	sig;
 
 	set_errno(EXIT_SUCCESS);
 	ms->input = readline("minishell$ ");
+	add_history(ms->input);
 	ms->input_line++;
 	sig = get_and_reset_signal();
 	if (sig == SIGINT)
@@ -108,7 +77,7 @@ int	main(int ac, char **av, char **envp)
 
 	(void)av;
 	if (ac > 1)
-		return (EXIT_FAILURE); //TODO Deal with non-interactive mode
+		return (put_error("no arguments allowed."), EXIT_FAILURE);
 	ft_signal(SIGQUIT, SIG_IGN);
 	ft_signal(SIGINT, rl_sigint_handler);
 	minishell = init_minishell(envp);
@@ -120,8 +89,8 @@ int	main(int ac, char **av, char **envp)
 			continue ;
 		ft_signal(SIGINT, exec_sigint_handler);
 		execute_cmd_lst(minishell);
-		ft_signal(SIGINT, rl_sigint_handler);
 		free_cmd_lst(&minishell->cmd_lst);
+		ft_signal(SIGINT, rl_sigint_handler);
 	}
 	free_minishell(&minishell);
 	return (EXIT_SUCCESS);
