+++ src/parser/cmd/new_cmd.c
@@ -1,9 +1,17 @@
 #include "minishell.h"
 
+#include "minishell.h"
+
+/**
+ * Initializes process-related fields in the t_cmd structure.
+ *
+ * @param cmd The command structure to initialize.
+ * @return EXIT_SUCCESS on success, EXIT_FAILURE on malloc failure.
+ */
 static int	init_cmd_process(t_cmd *cmd)
 {
 	cmd->pid = -1;
-	cmd->pipe = malloc(sizeof(int) *2);
+	cmd->pipe = malloc(sizeof(int) * 2);
 	if (!cmd->pipe)
 		return (EXIT_FAILURE);
 	cmd->pipe[0] = -1;
@@ -13,6 +21,12 @@ static int	init_cmd_process(t_cmd *cmd)
 	return (EXIT_SUCCESS);
 }
 
+/**
+ * Initializes heredoc-related fields in the t_cmd structure.
+ *
+ * @param cmd The command structure to initialize.
+ * @return EXIT_SUCCESS on success, EXIT_FAILURE on malloc failure.
+ */
 static int	init_cmd_heredoc(t_cmd *cmd)
 {
 	cmd->heredoc = malloc(sizeof(t_heredoc));
@@ -24,9 +38,13 @@ static int	init_cmd_heredoc(t_cmd *cmd)
 	return (EXIT_SUCCESS);
 }
 
-/*
-* Initializes redirection-related fields in the t_cmd structure.
-*/
+/**
+ * Initializes redirection-related fields in the t_cmd structure.
+ *
+ * @param cmd The command structure to initialize.
+ * @return EXIT_SUCCESS on success, EXIT_FAILURE on heredoc initialization
+ * failure.
+ */
 static int	init_cmd_redirections(t_cmd *cmd)
 {
 	cmd->fdin = STDIN_FILENO;
@@ -39,10 +57,13 @@ static int	init_cmd_redirections(t_cmd *cmd)
 	return (EXIT_SUCCESS);
 }
 
-/*
-* Creates a new command structure with default values.
-* Returns: A new t_cmd structure or NULL if malloc fails.
-*/
+/**
+ * Creates a new command structure with default values.
+ *
+ * @param prev_cmd The previous command in the list, or NULL if this is the
+ * first command.
+ * @return A new t_cmd structure on success, NULL on malloc failure.
+ */
 t_cmd	*cmd_new(t_cmd *prev_cmd)
 {
 	t_cmd	*cmd;
