+++ src/parser/handlers/handle_pipe.c
@@ -20,8 +20,6 @@ static char	*pipe_error(t_token *token)
  * Handles the pipe (`|`) token by creating a new command.
  * - Checks for syntax errors like consecutive pipes.
  * - Creates a new command for the next part of the pipeline.
- * @TODO Check malloc error message (for now on NULL)
- * 
  */
 int	handle_pipe(t_token **cur_token, t_cmd **cur_cmd,
 	t_minishell *minishell)
@@ -38,7 +36,7 @@ int	handle_pipe(t_token **cur_token, t_cmd **cur_cmd,
 	}
 	(*cur_cmd)->next = cmd_new((*cur_cmd));
 	if (!(*cur_cmd)->next)
-		exit_minishell(EXIT_FAILURE, minishell, NULL);
+		exit_minishell(EXIT_FAILURE, minishell, "cmd: malloc");
 	(*cur_cmd) = (*cur_cmd)->next;
 	*cur_token = token->next;
 	return (EXIT_SUCCESS);
