+++ src/parser/parser.c
@@ -1,20 +1,16 @@
 #include "minishell.h"
+
 /**
- * Cleans up token and command lists if parsing fails.
+ * Cleans up token and command lists.
  * 
  * @param minishell Shell instance containing token/cmd lists.
  * @param status Indicates failure (EXIT_FAILURE) or success.
  * @return The input `status` unchanged.
- * @note Frees resources only on failure (EXIT_FAILURE).
  */
-static int	parse_cleanup(t_minishell *minishell, int status)
+static void	parse_cleanup(t_minishell *minishell)
 {
-	if (status == EXIT_FAILURE)
-	{
-		free_token_lst(&minishell->token_lst);
-		free_cmd_lst(&minishell->cmd_lst);
-	}
-	return (status);
+	free_token_lst(&minishell->token_lst);
+	free_cmd_lst(&minishell->cmd_lst);
 }
 
 /* 
@@ -30,21 +26,17 @@ int	parse_tokens(t_minishell *minishell)
 {
 	t_token	*cur_token;
 	t_cmd	*cur_cmd;
-	t_token	*prev_token;
 
 	minishell->cmd_lst = cmd_new(NULL);
 	if (!minishell->cmd_lst)
-		return (parse_cleanup(minishell, EXIT_FAILURE));
+		return (parse_cleanup(minishell), EXIT_FAILURE);
 	cur_token = minishell->token_lst;
 	cur_cmd = minishell->cmd_lst;
 	while (cur_token)
 	{
-		prev_token = cur_token;
 		if (handle_token_type(&cur_token, &cur_cmd, minishell)
 			== EXIT_FAILURE)
-			return (parse_cleanup(minishell, EXIT_FAILURE));
-		if (cur_token == prev_token)
-			exit_minishell(EXIT_FAILURE, minishell, NULL);
+			return (parse_cleanup(minishell), EXIT_FAILURE);
 	}
 	return (EXIT_SUCCESS);
 }
