+++ src/signal/signal.c
@@ -2,24 +2,33 @@
 
 static volatile sig_atomic_t	g_signal = 0;
 
+/**
+ * Retrieves and resets the last received signal.
+ */
 int	get_and_reset_signal(void)
 {
 	sig_atomic_t	sig;
-	
+
 	sig = g_signal;
 	g_signal = 0;
 	return (sig);
 }
 
+/**
+ * Handles SIGINT during minishell input (main readline).
+ */
 void	rl_sigint_handler(int signal)
 {
 	g_signal = signal;
 	ft_printf("\n");
 	rl_replace_line("", 0);
 	rl_on_new_line();
-	rl_redisplay(); //TODO delete line?
+	rl_redisplay();
 }
 
+/**
+ * Handles SIGINT during heredoc input (sub-readline).
+ */
 void	heredoc_sigint_handler(int signal)
 {
 	g_signal = signal;
@@ -30,7 +39,11 @@ void	heredoc_sigint_handler(int signal)
 	close(STDIN_FILENO);
 }
 
+/**
+ * Handles SIGINT during command execution.
+ */
 void	exec_sigint_handler(int signal)
 {
 	g_signal = signal;
+	ft_printf("\n");
 }
