+++ src/tokenizer/tokenizer_utils.c
@@ -1,10 +1,10 @@
 #include "minishell.h"
-/* 
+/**
  * Returns a static list of tokenization operators and their corresponding types.
  * 
  * Used in `handle_special_char()` to identify and categorize operators.
  * 
- * Returns: A pointer to a static array of `t_tokenize_op` structures.
+ * @return Pointer to a static array of `t_tokenize_op` structures.
  */
 t_tokenize_op	*get_tokenize_ops(void)
 {
@@ -20,14 +20,16 @@ t_tokenize_op	*get_tokenize_ops(void)
 	return (tokenize_ops);
 }
 
-/* 
+/**
  * Handles special characters (`|`, `<`, `>`, `<<`, `>>`).
  *
  * - Compares the input string with predefined token patterns.
  * - Creates a token of the corresponding type if a match is found.
  * - Updates the input index accordingly.
  * 
- * Returns: A new token (`t_token *`) or NULL if no match is found.
+ * @param input The input string to parse.
+ * @param i Pointer to the current index in the input string.
+ * @returns New token (`t_token *`) or NULL if no match is found.
  */
 t_token	*handle_special_char(char *input, int *i)
 {
@@ -52,13 +54,16 @@ t_token	*handle_special_char(char *input, int *i)
 	return (NULL);
 }
 
-/* 
+/**
  * Determines whether the next token is a special character or a word.
  *
  * - Calls `handle_special_char()` if the character is a special token.
  * - Calls `create_word_token()` otherwise.
  * 
- * Returns: A newly allocated token (`t_token *`).
+ * @param input The input string to parse.
+ * @param i Pointer to the current index in the input string.
+ * @param unmatched_quote Pointer to the unmatched quote character, if any.
+ * @returns Newly allocated token (`t_token *`).
  */
 t_token	*handle_token_creation(char *input, int *i, char *unmatched_quote)
 {
@@ -67,13 +72,16 @@ t_token	*handle_token_creation(char *input, int *i, char *unmatched_quote)
 	return (create_word_token(input, i, unmatched_quote));
 }
 
-/* 
+/**
  * Handles tokenization errors, such as unmatched quotes.
  * 
  * - Frees the token list if an error is encountered.
  * - Exits the minishell with an appropriate error message.
  * 
- * Returns: Always returns 0.
+ * @param token_lst Pointer to the list of tokens to free.
+ * @param unmatched_quote The unmatched quote character, if any.
+ * @param minishell Pointer to the minishell data structure.
+ * @return Always returns EXIT_SUCCESS.
  * //TODO Interactive mode in case of unmatching quotes?
  */
 int	handle_token_error(t_token **token_lst, char unmatched_quote,
@@ -86,5 +94,5 @@ int	handle_token_error(t_token **token_lst, char unmatched_quote,
 			unmatched_quote);
 	else
 		exit_minishell(EXIT_FAILURE, minishell, NULL);
-	return (0);
+	return (EXIT_SUCCESS);
 }
